Better image drawing performance, .NET 6 and more! (#951)
* Better image drawing performance, .NET 6 and more!

* Update workflows

* Changes

* Separate .NET 6 changes from PR

* Update workflows to .NET 5

* Use correct operators

* Changes

* Fix CoolWorld not testing correctly

* Handle padding in Bitmap

* Fix error

* Update Bitmap logic

* Update Bitmap logic

* Use do-while loop instead of for loop

* Remove .idea folder

* Update Bitmap logic

+++ Source/Mosa.DeviceSystem/ByteStream.cs
@@ -11,24 +11,24 @@ namespace Mosa.DeviceSystem
 			this.data = data;
 		}
 
-		public byte Read8(int offset)
+		public byte Read8(uint offset)
 		{
 			return data[offset];
 		}
 
-		public int Read16(int offset)
+		public ushort Read16(uint offset)
 		{
-			return data[offset] | data[offset + 1] << 8;
+			return (ushort)(data[offset] | data[offset + 1] << 8);
 		}
 
-		public int Read24(int offset)
+		public uint Read24(uint offset)
 		{
-			return data[offset] | data[offset + 1] << 8 | data[offset + 2] << 16;
+			return (uint)(data[offset] | data[offset + 1] << 8 | data[offset + 2] << 16);
 		}
 
-		public int Read32(int offset)
+		public uint Read32(uint offset)
 		{
-			return data[offset] | data[offset + 1] << 8 | data[offset + 2] << 16 | data[offset + 3] << 24;
+			return (uint)(data[offset] | data[offset + 1] << 8 | data[offset + 2] << 16 | data[offset + 3] << 24);
 		}
 	}
 }
